feat(connectors): [BREAKING] mandatory `to` or `from` options for a node.
Specify `to`, `from` options to draw a connector to or from a node to the node that has this option set. They can either be array of ids(string) or objects of type `{id: string, path: string}`

If no `to` or `from` option is provided, no connector is drawn from or to that node.